Fix hovering on base of center laptop
Before, the center laptop could not be hovered on the flat top of its
base. This is because only the .wrapper elements are transformed,
leaving all of the .laptop3d elements in the center middle (even though
nothing is rendered there). Regardless, they were catching hover events.
By applying pointer-events: none and re-enabling pointer-events for the
.wrapper element, this is resolved.